Updated some relationships and specifically the OnDelete action

+++ Akla.SharedData/Models/Order.cs
@@ -9,8 +9,8 @@
         [MaxLength(255)]
         public string Delivery_Address { get; set; }
 
-        public int PaymentMethod_Id { get; set; }
-        public int Customer_Id { get; set; }
+        public long PaymentMethod_Id { get; set; }
+        public long Customer_Id { get; set; }
 
         public virtual PaymentMethod PaymentMethod { get; set; }
         public virtual Customer Customer { get; set; }
